# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "StoreFy API",
        "description": "StoreFY Swagger OpenApi description",
        "contact": {
            "email": "georgeton.ufg@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Api local server"
        },
        {
            "url": "http://storefy.workingtech.com.br:31080/api",
            "description": "Api dev server"
        }
    ],
    "paths": {
...
        "/cfg-boletos": {
            "get": {
                "tags": [
                    "cfg-boletos"
                ],
                "summary": "Get all items",
                "operationId": "getCfgBoletos",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/fields"
                    },
                    {
                        "$ref": "#/components/parameters/where"
                    },
                    {
                        "$ref": "#/components/parameters/orderBy"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "title": "id",
                                                    "description": "id",
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/CfgBoletos"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "cfg-boletos"
                ],
                "summary": "Add a new item",
                "operationId": "addCfgBoletos",
                "requestBody": {
                    "description": "Object to add new item",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CfgBoletos"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "title": "id",
                                                    "description": "id",
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/CfgBoletos"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "$ref": "#/components/schemas/validate_message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cfg-boletos/{uuid}": {
            "get": {
                "tags": [
                    "cfg-boletos"
                ],
                "summary": "Get item by ID",
                "operationId": "getCfgBoletosById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "$ref": "#/components/parameters/fields"
                    },
                    {
                        "$ref": "#/components/parameters/where"
                    },
                    {
                        "$ref": "#/components/parameters/orderBy"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "title": "id",
                                                    "description": "id",
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/CfgBoletos"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "cfg-boletos"
                ],
                "summary": "Update item",
                "operationId": "updateCfgBoletos",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Object to update item",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CfgBoletos"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "title": "id",
                                                    "description": "id",
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/CfgBoletos"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "cfg-boletos"
                ],
                "summary": "Delete item",
                "operationId": "deleteCfgBoletos",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Item not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "validate_message": {
                "title": "Error message",
                "description": "Error jso1n",
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "type": "string",
                        "readOnly": true
                    },
                    "message": {
                        "type": "string",
                        "minLength": 8,
                        "readOnly": true
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ErrorMessage"
                }
            },
...
            "CfgBoletos": {
                "title": "CfgBoletos model",
                "description": "CfgBoletos model",
                "required": [],
                "properties": {
                    "field name": {
                        "title": "field title",
                        "description": "Field name"
                    },
                    "status": {
                        "title": "status",
                        "description": "Status",
                        "type": "boolean",
                        "default": false
                    },
                    "uuid": {
                        "title": "uuid",
                        "description": "Uuid"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CfgBoletos"
                }
            },
...
        },
        "parameters": {
            "fields": {
                "name": "fields",
                "in": "query",
                "description": "Fields to return"
            },
            "where": {
                "name": "where",
                "in": "query",
                "description": "filter of any database fields (always send ternaries: 'field,operator,value')'"
            },
            "orderBy": {
                "name": "orderBy",
                "in": "query",
                "description": "Order ny database fields (always send: 'field,order')'"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}